requirements.txt
flask==3.0.0
flask-cors==4.0.0
tflite-runtime==2.14.0
Pillow==10.4.0
numpy==1.26.4
requests==2.31.0
gunicorn==21.2.0
google-generativeai==0.8.3
